Add Note Edit view specific to CalendarDisplayView to avoid changing to NotesDisplayView when editing. Remove margins from content views.
