+++ PlanSimple/App.xaml
@@ -33,6 +33,12 @@
              <DataTemplate DataType="{x:Type viewModel:CalendarViewModel}">
                  <view:CalendarView/>
              </DataTemplate>
+             <DataTemplate DataType="{x:Type viewModel:CalendarDisplayViewModel}">
+                 <view:CalendarDisplayView/>
+             </DataTemplate>
+             <DataTemplate DataType="{x:Type viewModel:CalendarNoteEditViewModel}">
+                 <view:CalendarNoteEditView/>
+             </DataTemplate>
              
              <DataTemplate DataType="{x:Type viewModel:NotesViewModel}">
                  <view:NotesView/>
