+++ PlanSimple/MVVM/View/CalendarDisplayView.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Controls;
+
+namespace PlanSimple.MVVM.View;
+
+public partial class CalendarDisplayView : UserControl
+{
+	public CalendarDisplayView()
+	{
+		InitializeComponent();
+	}
+}
