+++ PlanSimple/MVVM/View/CalendarNoteEditView.xaml
@@ -1,0 +1,106 @@
+﻿<UserControl x:Class="PlanSimple.MVVM.View.CalendarNoteEditView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:i="http://schemas.microsoft.com/xaml/behaviors"
+             xmlns:viewModel="clr-namespace:PlanSimple.MVVM.ViewModel"
+             xmlns:view="clr-namespace:PlanSimple.MVVM.View"
+             xmlns:controls="clr-namespace:PlanSimple.Controls"
+             xmlns:helpers="clr-namespace:PlanSimple.HelperClasses"
+             mc:Ignorable="d"
+             d:DesignHeight="300" d:DesignWidth="300">
+    <UserControl.DataContext>
+        <viewModel:CalendarNoteEditViewModel/>
+    </UserControl.DataContext>
+    
+    <Grid Background="{DynamicResource Background}">
+        <Grid.RowDefinitions>
+            <RowDefinition/>
+            <RowDefinition Height="100"/>
+        </Grid.RowDefinitions>
+        
+        <Grid Grid.Row="0"
+              Margin="30">
+            <Grid.RowDefinitions>
+                <RowDefinition Height="100"/>
+                <RowDefinition/>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition Height="Auto"/>
+            </Grid.RowDefinitions>
+            
+            <TextBox Grid.Row="0"
+                     Tag="Note Title..."
+                     Style="{StaticResource TextboxWithPlaceholder}"
+                     Text="{Binding ToDoNote.Title}"/>
+            
+            <TextBox Grid.Row="1"
+                     Tag="Note Description..."
+                     Style="{StaticResource TextboxWithPlaceholder}"
+                     Text="{Binding ToDoNote.Description}"/>
+            
+            <Grid Grid.Row="2"
+                  Margin="1,5,1,0">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="100"/>
+                    <ColumnDefinition/>
+                </Grid.ColumnDefinitions>
+                <Label Grid.Column="0" Content="Priority"/>
+                
+                <ComboBox Grid.Column="1"
+                          Padding="100,0,0,0"
+                          HorizontalAlignment="Stretch"
+                          VerticalAlignment="Stretch"
+                          ItemsSource="{Binding ToDoNote.Priority, Converter={helpers:EnumToCollectionConverter}, Mode=OneTime}"
+                          SelectedValuePath="Item1"
+                          DisplayMemberPath="Item2"
+                          SelectedValue="{Binding ToDoNote.Priority}"
+                          Foreground="{DynamicResource Foreground}" />
+            </Grid>
+            
+            <DatePicker Grid.Row="3"
+                        Margin="1,5,1,0"
+                        Style="{StaticResource DatePickerWithPlaceholder}"
+                        SelectedDate="{Binding Date}"/>
+        </Grid>
+        
+        <Grid Grid.Row="1">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition/>
+                <ColumnDefinition/>
+            </Grid.ColumnDefinitions>
+            
+            <controls:HoverButtonWithIcon Grid.Column="0"
+                                          Background="{DynamicResource Background-3}"
+                                          HoverBackground="{DynamicResource Background-10}"
+                                          Foreground="{DynamicResource Foreground}"
+                                          HoverForeground="{DynamicResource Foreground}"
+                                          BorderThickness="0"
+                                          Content="Save"
+                                          FontSize="22"
+                                          Icon="&#xE82E;"
+                                          IconFontFamily="Segoe MDL2 Assets"
+                                          Command="{Binding SaveDataToDatabase}">
+                <i:Interaction.Triggers>
+                    <i:EventTrigger EventName="Click">
+                        <i:InvokeCommandAction Command="{Binding SaveDataToDatabase}"/>
+                        <i:InvokeCommandAction Command="{Binding DataContext.CalendarDisplayViewCommand,
+                          RelativeSource={RelativeSource AncestorType={x:Type view:CalendarView}}}"/>
+                    </i:EventTrigger>
+                </i:Interaction.Triggers>
+            </controls:HoverButtonWithIcon>
+
+            <controls:HoverButtonWithIcon Grid.Column="1"
+                                          Background="{DynamicResource Background-3}"
+                                          HoverBackground="{DynamicResource Background-10}"
+                                          Foreground="{DynamicResource Foreground}"
+                                          HoverForeground="{DynamicResource Foreground}"
+                                          BorderThickness="0"
+                                          Content="Cancel"
+                                          FontSize="22"
+                                          Icon="&#xF78A;"
+                                          IconFontFamily="Segoe MDL2 Assets"
+                                          Command="{Binding DataContext.CalendarDisplayViewCommand, RelativeSource={RelativeSource AncestorType={x:Type view:CalendarView}}}" />
+        </Grid>
+    </Grid>
+</UserControl>
