+++ PlanSimple/MVVM/View/CalendarNoteEditView.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Controls;
+
+namespace PlanSimple.MVVM.View;
+
+public partial class CalendarNoteEditView : UserControl
+{
+	public CalendarNoteEditView()
+	{
+		InitializeComponent();
+	}
+}
