+++ PlanSimple/MVVM/View/CalendarView.xaml.cs
@@ -1,33 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
+﻿using System.Windows.Controls;
 
-namespace PlanSimple.MVVM.View
+namespace PlanSimple.MVVM.View;
+
+/// <summary>
+/// Interaction logic for CalendarView.xaml
+/// </summary>
+public partial class CalendarView : UserControl
 {
-    /// <summary>
-    /// Interaction logic for CalendarView.xaml
-    /// </summary>
-    public partial class CalendarView : UserControl
+    public CalendarView()
     {
-        public CalendarView()
-        {
-            InitializeComponent();
-        }
-
-        private void HoverButton_Click(object sender, RoutedEventArgs e)
-        {
-
-        }
+        InitializeComponent();
     }
 }
